+++ make_responsive_images/main.py
@@ -42,19 +42,58 @@ def image(
             .parent.parent.joinpath("tests")
             .joinpath("fixtures")
             .joinpath("xfer-original.jpg")
-        )
+        ),
+        help="Image file location"
     ),
-    widths: str = typer.Option("600,1000,1400"),
+    widths: str = typer.Option(
+        "600,1000,1400",
+        help="Widths of new images, in pixels"
+    ),
+    html: bool = typer.Option(
+        True,
+        help="Generate HTML <img> tag"
+    ),
+    classes: str = typer.Option(
+        None,
+        help='Classnames to add to the <img> tag (e.g. class="img-fluid")'
+    ),
+    img_sizes: str = typer.Option(
+        "100vw",
+        help='Sizes for the <img> tag (e.g. sizes="100vw")'
+    ),
+    lazy: bool = typer.Option(
+        True,
+        help='Adds loading="lazy" to <img> tag for SEO'
+    ),
+    alt: str = typer.Option(
+        "",
+        help='Adds alt="" to the <img> tag (e.g. alt="Funny image")'
+    ),
+    dir: str = typer.Option(
+        None,
+        help='Images directory to prepend to the src (e.g. src="<dir>/<image>")'
+    ),
 ) -> None:
     """Resize one image"""
 
     typer.secho(f"Image: {image}", fg=typer.colors.GREEN)
-    typer.secho(f"Widths needed: {widths}", fg=typer.colors.GREEN)
+    typer.echo(f"Widths needed: {widths}")
+    typer.echo(f"HTML wanted: {html}")
 
     widths_split = widths.split(",")
     widths_list = [int(width) for width in widths_split]
 
     file = Path(image)
-    images = resize_image(file, widths=widths_list)
+    filenames = resize_image(
+        file=file,
+        widths=widths_list,
+        html=html,
+        classes=classes,
+        img_sizes=img_sizes,
+        lazy=lazy,
+        alt=alt,
+        dir=dir,
+    )
+    typer.echo(f"filenames: {filenames}")
 
-    typer.secho(f"Images: {images}", fg=typer.colors.GREEN)
+