+++ make_responsive_images/utils.py
@@ -3,7 +3,16 @@ from pathlib import Path
 from PIL import Image
 
 
-def resize_image(file: Path, widths: list) -> list:
+def resize_image(
+    file: Path,
+    widths: list,
+    html: bool,
+    classes: str,
+    img_sizes: str,
+    lazy: bool,
+    alt: str,
+    dir: str,
+) -> list:
     """Resize the image to the widths specified"""
 
     image = Image.open(file)
@@ -25,6 +34,7 @@ def resize_image(file: Path, widths: list) -> list:
 
     # create the resized images
     resized = []
+    filenames = []
     for (width, height) in sizes:
         if (width, height) == (image.width, image.height):
             resized.append(image)
@@ -51,4 +61,45 @@ def resize_image(file: Path, widths: list) -> list:
         path_new = file.parent.joinpath(filename_new)
         new_image.save(path_new)
 
-    return resized
+        filenames.append((filename_new, width))
+
+    if html:
+        html_str = '<img '
+
+        if classes:
+            html_str += f'class="{classes}" '
+        if lazy:
+            html_str += 'loading="lazy" '
+
+        html_str += f'\n  sizes="{img_sizes}" '
+        html_str += f'\n  alt="{alt}" '
+        
+        def _get_filename(dir, filename):
+            if dir:
+                return f"{dir}/{filename}"
+            return filename
+
+        src_filename = _get_filename(dir, file.name)
+        html_str += f'\n  src="{src_filename}" '
+
+        srcset_str = '\n  srcset="'
+        n_filenames = len(filenames)
+        for num, (filename, width) in enumerate(filenames):
+            is_last = num == (n_filenames - 1)
+                # {{ url_for('static', filename='img/vru/ijack-egas-vru-sizes-600px.jpg') }} 600w,
+                # {{ url_for('static', filename='img/vru/ijack-egas-vru-sizes-1000px.jpg') }} 1000w,
+                # {{ url_for('static', filename='img/vru/ijack-egas-vru-sizes-1600px.jpg') }} 1600w"
+            filename = _get_filename(dir, filename)
+            if is_last:
+                srcset_str += f'\n    {filename} {width}w"'
+            else:
+                srcset_str += f'\n    {filename} {width}w,'
+        
+        html_str += srcset_str
+        html_str += ">"
+        
+        html_path = file.parent.joinpath("img_tag.html")
+        with open(html_path, "w") as f:
+            f.write(html_str)
+
+    return filenames