+++ CHANGELOG.md
@@ -13,3 +13,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [Unreleased]: https://github.com/mccarthysean/make-responsive-images/compare/v0.1.1...HEAD
 [0.1.1]: https://github.com/mccarthysean/make-responsive-images/releases/tag/v0.1.1
+
+## [v0.1.2]
+
+## [0.1.2] - 2021-11-13
+
+- First working version, which resizes one image file
+
+[v0.1.2]: https://github.com/mccarthysean/make-responsive-images/compare/v0.1.2...HEAD
+[0.1.2]: https://github.com/mccarthysean/make-responsive-images/releases/tag/v0.1.2
+
+## [v0.1.3]
+
+## [0.1.3] - 2021-11-14
+
+- More `<img>` tag options, and now it generates an "img_tag.html" file with the new `<img>` tag.
+
+[v0.1.3]: https://github.com/mccarthysean/make-responsive-images/compare/v0.1.3...HEAD
+[0.1.3]: https://github.com/mccarthysean/make-responsive-images/releases/tag/v0.1.3